requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
streamlit
python-dotenv